+++ context/app/static/js/components/cells/LogSliderWrapper.js
